+++ scripts/release.sh
@@ -2,6 +2,6 @@
 set -euv
 
 sh -c "NODE_ENV=production yarn build"
-cp -r LICENSE README.md package.json typings built lib/
+cp -r LICENSE README.md package.json typings lib/
 cp examples/basic.js lib/example.js
 sh -c "cd lib; npm publish"