+++ spec/integration/browser/karma.conf.js
@@ -6,43 +6,38 @@ module.exports = function (config) {
 
   config.set({
     browsers: process.platform === 'win32' ? ['EdgeHeadless'] : ['ChromeHeadless'],
-    frameworks: ['jasmine', 'karma-typescript'],
-    reporters: ['spec', 'karma-typescript'],
+    frameworks: ['jasmine'],
+    reporters: ['spec'],
 
     plugins: [
       'karma-webpack',
       'karma-jasmine',
-      'karma-typescript',
       'karma-sourcemap-loader',
       'karma-chrome-launcher',
       '@chiragrupani/karma-chromium-edge-launcher',
-      'karma-spec-reporter'
+      'karma-spec-reporter',
     ],
 
     singleRun: process.env.SINGLE_RUN || false,
 
-    files: [
-      { pattern: '../../../src/**/*.ts', watched: false },
-      { pattern: '*.spec.js', watched: false }
-    ],
+    files: [{ pattern: '*.spec.js', watched: false }],
 
     preprocessors: {
       '*.spec.js': ['webpack', 'sourcemap'],
-      '../../../src/**/*.ts': ['karma-typescript', 'webpack']
     },
 
     proxies: {
-      '/proxy': 'http://localhost:9090'
+      '/proxy': 'http://localhost:9090',
     },
 
     karmaTypescriptConfig: {
-      tsconfig: '../../../tsconfig.karma.json'
+      tsconfig: '../../../tsconfig.karma.json',
     },
 
     webpack: webpackConfig,
 
     webpackMiddleware: {
-      stats: 'errors-only'
-    }
+      stats: 'errors-only',
+    },
   })
 }